Refactoring and enhancements
Rename gen-cloud-init-iso.py to gen_cloud_init_iso.py

use sys.exit(1) instead of sys.exit()

use ignore_errors=True instead of IF/action

fix resource block name inconsistencies

use helper functions

remove cloud-init isos when destroying
