+++ ci/infra/vmware/cloud-init.tf
@@ -1,12 +1,20 @@
-data "local_file" "cloud-init-metadata" {
+data "local_file" "cloud_init_metadata" {
     filename = "cloud-init/meta-data.ccfile"
 }
 
-data "local_file" "cloud-init-netconfig" {
+data "local_file" "cloud_init_netconfig" {
     filename = "cloud-init/network-config.ccfile"
 }
 
 locals {
-  cloud-init-metadata = "${data.local_file.cloud-init-metadata.content}"
-  cloud-init-netconfig = "${data.local_file.cloud-init-netconfig.content}"
+  cloud_init_metadata  = "${data.local_file.cloud_init_metadata.content}"
+  cloud_init_netconfig = "${data.local_file.cloud_init_netconfig.content}"
 }
+
+resource "null_resource" "local_clean_up_isos" {
+  provisioner "local-exec" {
+    when    = "destroy"
+    command = "rm -f $PWD/cc-*.iso"
+  }
+}
+
