+++ ci/infra/vmware/master-instance.tf
@@ -8,7 +8,7 @@ data "template_file" "master_repositories" {
   }
 }
 
-data "template_file" "master_cloud-init-userdata" {
+data "template_file" "master_cloud_init_userdata" {
   template = "${file("cloud-init/master.tpl")}"
 
   vars {
@@ -20,9 +20,9 @@ data "template_file" "master_cloud-init-userdata" {
   }
 }
 
-resource "null_resource" "local_gen-cc-master-iso" {
+resource "null_resource" "local_gen_cc_master_iso" {
   provisioner "local-exec" {
-    command = "./gen-cloud-init-iso.py master '${data.template_file.master_cloud-init-userdata.rendered}' '${local.cloud-init-metadata}' '${local.cloud-init-netconfig}'"
+    command = "./gen_cloud_init_iso.py master '${data.template_file.master_cloud_init_userdata.rendered}' '${local.cloud_init_metadata}' '${local.cloud_init_netconfig}'"
   }
 }
 
@@ -32,7 +32,7 @@ resource "vsphere_file" "upload_cc_master_iso" {
   source_file      = "./cc-master.iso"
   create_directories = true
   destination_file = "${var.stack_name}/cc-master.iso"
-  depends_on = ["null_resource.local_gen-cc-master-iso"]
+  depends_on = ["null_resource.local_gen_cc_master_iso"]
 }
 
 resource "vsphere_virtual_machine" "master" {
